% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_infosiga.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get Infosiga Cache Directory}
\usage{
get_cache_dir(package = "infosiga")
}
\arguments{
\item{package}{Character. Name of the package. Defaults to "infosiga".}
}
\value{
A character string with the path to the cache directory.
}
\description{
Returns the path to the user-specific cache directory used to store
downloaded Infosiga dataset files.
}
\examples{
\dontrun{
# Get the path to the cache directory
get_cache_dir()
}

}
