% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infosiga_sinistros}
\alias{infosiga_sinistros}
\title{Road Crashes in São Paulo}
\format{
\subsection{\code{infosiga_sinistros}}{

A \code{tibble} with 797,575 observations and 36 variables:
\describe{
\item{id_sinistro}{Unique identifier for the occurrence. (numeric)}
\item{data_sinistro}{Date of the road crash. (class \code{Date})}
\item{hora_sinistro}{Time of the road crash. (class \code{hms})}
\item{cod_ibge}{IBGE code of the municipality where the road crash occurred. (numeric)}
\item{municipio}{Name of the municipality where the road crash occurred. (character)}
\item{numero_logradouro}{Street number where the road crash occurred. (numeric)}
\item{tipo_via}{Type of road where the road crash occurred. (character)}
\item{longitude}{Longitude of the road crash location. (numeric)}
\item{latitude}{Latitude of the road crash location. (numeric)}
\item{tp_veiculo_bicicleta}{Number of bicycles involved in the road crash. (numeric)}
\item{tp_veiculo_caminhao}{Number of trucks involved in the road crash. (numeric)}
\item{tp_veiculo_motocicleta}{Number of motorcycles involved in the road crash. (numeric)}
\item{tp_veiculo_nao_disponivel}{Number of vehicles with unavailable classification. (numeric)}
\item{tp_veiculo_onibus}{Number of buses involved in the road crash. (numeric)}
\item{tp_veiculo_outros}{Number of other types of vehicles involved. (numeric)}
\item{tp_veiculo_automovel}{Number of cars involved in the road crash. (numeric)}
\item{gravidade_nao_disponivel}{Road crashes with unknown severity. (numeric)}
\item{gravidade_leve}{Road crashes with minor injuries. (numeric)}
\item{gravidade_fatal}{Fatal road crashes. (numeric)}
\item{gravidade_ileso}{Road crashes where individuals were uninjured. (numeric)}
\item{gravidade_grave}{Road crashes with serious injuries. (numeric)}
\item{administracao_via}{Entity responsible for road administration. (character)}
\item{jurisdicao_via}{Jurisdiction of the road where the road crash occurred. (character)}
\item{tipo_sinistro_primario}{Primary type of road crash. (character)}
\item{tp_sinistro_atropelamento}{Pedestrian-related crashes. (numeric)}
\item{tp_sinistro_colisao_frontal}{Head-on collisions. (numeric)}
\item{tp_sinistro_colisao_traseira}{Rear-end collisions. (numeric)}
\item{tp_sinistro_colisao_lateral}{Side collisions. (numeric)}
\item{tp_sinistro_colisao_transversal}{Cross collisions. (numeric)}
\item{tp_sinistro_colisao_outros}{Other types of collisions. (numeric)}
\item{tp_sinistro_choque}{Shock-related crashes. (numeric)}
\item{tp_sinistro_capotamento}{Rollover crashes. (numeric)}
\item{tp_sinistro_engavetamento}{Pile-up crashes. (numeric)}
\item{tp_sinistro_tombamento}{Overturn crashes. (numeric)}
\item{tp_sinistro_outros}{Other types of crashes. (numeric)}
\item{tp_sinistro_nao_disponivel}{Crashes with an unspecified type. (numeric)}
}
}
}
\source{
Infosiga.SP
}
\usage{
infosiga_sinistros
}
\description{
This dataset contains detailed information about road crashes that
occurred in the state of São Paulo, sourced from Infosiga.SP.
}
\references{
\url{https://www.infosiga.sp.gov.br/}
}
\keyword{datasets}
