% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infosiga_veiculos}
\alias{infosiga_veiculos}
\title{Vehicles Involved in Road Crashes in São Paulo}
\format{
A \code{tibble} with 1,364,537 observations and 5 variables:
\describe{
\item{id_sinistro}{Unique identifier for the road crash. (numeric)}
\item{ano_fabricacao}{Year the vehicle was manufactured. (numeric)}
\item{ano_modelo}{Model year of the vehicle. (numeric)}
\item{cor_veiculo}{Color of the vehicle. (character)}
\item{tipo_veiculo}{Type of vehicle (character)}
}
}
\source{
Infosiga.SP
}
\usage{
infosiga_veiculos
}
\description{
This dataset contains information about vehicles involved in road crashes
in São Paulo, including their manufacturing year, model year,
color, and type.
}
\references{
\url{https://www.infosiga.sp.gov.br/}
}
\keyword{datasets}
