% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infosiga_vitimas}
\alias{infosiga_vitimas}
\title{Road Crash Victims in São Paulo}
\format{
\subsection{\code{infosiga_vitimas}}{

A \code{tibble} with 1,618,002 observations and 9 variables:
\describe{
\item{id_sinistro}{Unique identifier for the road crash. (numeric)}
\item{data_sinistro}{Date of the road crash. (class \code{Date})}
\item{data_obito}{Date of death, if applicable. (class \code{Date})}
\item{sexo}{Gender of the victim. (character)}
\item{idade}{Age of the victim. (numeric)}
\item{tipo_vitima}{Type of victim (character)}
\item{faixa_etaria}{Age group of the victim.}
\item{tipo_veiculo_vitima}{Type of vehicle the victim was using (character)}
\item{gravidade_lesao}{Injury severity (character)}
}
}
}
\source{
Infosiga.SP
}
\usage{
infosiga_vitimas
}
\description{
This dataset contains information about individuals involved
in road crashes in the state of São Paulo, including their demographics,
injury severity, and vehicle type.
}
\references{
\url{https://www.infosiga.sp.gov.br/}
}
\keyword{datasets}
