% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_infosiga.R
\name{load_infosiga}
\alias{load_infosiga}
\title{Load data from Infosiga.SP}
\usage{
load_infosiga(type = c("sinistros", "vitimas", "veiculos"))
}
\arguments{
\item{type}{One of \code{"sinistros"}, \code{"vitimas"}, or \code{"veiculos"}. Determines which dataset to load.}
}
\value{
A \code{data.frame} with the selected dataset.
}
\description{
Downloads and loads a \code{.parquet} dataset (sinistros, vítimas ou veículos).
}
\examples{
\dontrun{
df <- load_infosiga("sinistros")
}

}
